+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 
+## Unreleased
+
+### 🩹 Fixes
+
+- **auth:** Prevent client overrides of reserved OAuth/OIDC authorization parameters
+
 ## v1.0.0-beta.11
 
 [compare changes](https://github.com/itpropro/nuxt-oidc-auth/compare/v1.0.0-beta.10...v1.0.0-beta.11)